requirements.txt
xlrd==2.0.1
pandas
requests
bs4
loguru
google-api-python-client
pygsheets
flask
flask-cors
gspread
line-bot-sdk==3.17.1
gunicorn
supabase
python-dotenv